requirements.txt
alembic==1.5.4
python-dotenv==0.15.0
is-safe-url==1.0
itsdangerous==1.1.0
email-validator==1.1.2
Click==7.1.2
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
SQLAlchemy==1.3.23
mysql-connector-python==8.0.23
Werkzeug==1.0.1
WTForms==2.3.3
gunicorn==20.0.4
Flask-Login==0.5.0
